Fix pedantic warning (size_t versus unsigned int)
git-svn-id: https://svn.science.uu.nl/repos/edu.3470784.Softwareproject_C@2 54bd1c4f-d025-4ee3-8271-c04213acb236

+++ jpeg_pyramid_builder/pyramid_builder.cpp
@@ -198,7 +198,7 @@ public:
     {
         if(size > 1)
         {
-            size_t csize = max(1u, size / 2);
+            size_t csize = max((size_t)1u, size / 2);
             Tile * subtiles [] =
             {
                       new Tile(this, x        , y        , z + 1, csize),
@@ -435,3 +435,4 @@ void processImage(const string & image_path, const string & output_pr,
         throw;
     }
 }
+
